Split ethnicity form selectors into their own file
